print list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,3 +1,4 @@
+#include <cstddef>
 #include <stdio.h>
 #include <stdlib.h>
 #include "lists.h"
@@ -11,7 +12,7 @@
 
 size_t print_list(const list_t *h)
 {
-	int count = 0;
+	size_t count = 0;
 	const list_t *current = NULL;
